updated naming to be same as sdk
