+++ server/api/contract/royalties/read/getDefaultRoyaltyInfo.ts
@@ -9,10 +9,8 @@ import { RoyaltySchema } from "../../../../schemas/contract";
 import { getChainIdFromChain } from "../../../../utilities/chain";
 import { getContract } from "../../../../utils/cache/getContract";
 
-const requestSchema = Type.Object({
-  tokenId: Type.String(),
-  ...contractParamSchema.properties,
-});
+const requestSchema = contractParamSchema;
+
 // OUTPUT
 const responseSchema = Type.Object({
   result: RoyaltySchema,
@@ -27,19 +25,19 @@ responseSchema.examples = [
   },
 ];
 
-export async function getTokenRoyaltyInfo(fastify: FastifyInstance) {
+export async function getDefaultRoyaltyInfo(fastify: FastifyInstance) {
   fastify.route<{
     Params: Static<typeof requestSchema>;
     Reply: Static<typeof responseSchema>;
   }>({
     method: "GET",
-    url: "/contract/:chain/:contract_address/royalty/get-token-royalty-info/:token_id",
+    url: "/contract/:chain/:contract_address/royalties/get-default-royalty-info",
     schema: {
-      summary: "Get Token Royalty Info",
+      summary: "Get Default Royalty Info",
       description:
-        "Gets the royalty recipient and BPS (basis points) of a particular token in the contract.",
-      tags: ["Contract-Royalty"],
-      operationId: "royalty_getDefaultRoyaltyInfo",
+        "Gets the royalty recipient and BPS (basis points) of the smart contract.",
+      tags: ["Contract-Royalties"],
+      operationId: "royalties_getDefaultRoyaltyInfo",
       params: requestSchema,
       response: {
         ...standardResponseSchema,
@@ -47,7 +45,7 @@ export async function getTokenRoyaltyInfo(fastify: FastifyInstance) {
       },
     },
     handler: async (request, reply) => {
-      const { chain, contract_address, tokenId } = request.params;
+      const { chain, contract_address } = request.params;
 
       const chainId = getChainIdFromChain(chain);
       const contract = await getContract({
@@ -55,7 +53,7 @@ export async function getTokenRoyaltyInfo(fastify: FastifyInstance) {
         contractAddress: contract_address,
       });
 
-      const returnData = await contract.royalties.getTokenRoyaltyInfo(tokenId);
+      const returnData = await contract.royalties.getDefaultRoyaltyInfo();
 
       reply.status(StatusCodes.OK).send({
         result: returnData,
